+++ packages/liveblocks-schema/src/ast/index.ts
@@ -28,7 +28,8 @@ export function isBuiltInScalarType(node: Node): node is BuiltInScalarType {
   return (
     node._kind === "StringKeyword" ||
     node._kind === "IntKeyword" ||
-    node._kind === "FloatKeyword"
+    node._kind === "FloatKeyword" ||
+    node._kind === "BooleanKeyword"
   );
 }
 
@@ -49,7 +50,11 @@ export function isTypeExpr(node: Node): node is TypeExpr {
   );
 }
 
-export type BuiltInScalarType = StringKeyword | IntKeyword | FloatKeyword;
+export type BuiltInScalarType =
+  | StringKeyword
+  | IntKeyword
+  | FloatKeyword
+  | BooleanKeyword;
 
 export type Definition = ObjectTypeDef;
 
@@ -64,6 +69,7 @@ export type TypeExpr =
 export type Range = [number, number];
 
 export type Node =
+  | BooleanKeyword
   | Document
   | FieldDef
   | FloatKeyword
@@ -73,6 +79,7 @@ export type Node =
   | ObjectLiteralExpr
   | ObjectTypeDef
   | StringKeyword
+  | TypeName
   | TypeRef;
 
 export function isRange(thing: unknown): thing is Range {
@@ -86,6 +93,7 @@ export function isRange(thing: unknown): thing is Range {
 
 export function isNode(node: Node): node is Node {
   return (
+    node._kind === "BooleanKeyword" ||
     node._kind === "Document" ||
     node._kind === "FieldDef" ||
     node._kind === "FloatKeyword" ||
@@ -95,9 +103,16 @@ export function isNode(node: Node): node is Node {
     node._kind === "ObjectLiteralExpr" ||
     node._kind === "ObjectTypeDef" ||
     node._kind === "StringKeyword" ||
+    node._kind === "TypeName" ||
     node._kind === "TypeRef"
   );
 }
+
+export type BooleanKeyword = {
+  _kind: "BooleanKeyword";
+  dummy_: string | null;
+  range: Range;
+};
 
 export type Document = {
   _kind: "Document";
@@ -145,7 +160,7 @@ export type ObjectLiteralExpr = {
 
 export type ObjectTypeDef = {
   _kind: "ObjectTypeDef";
-  name: Identifier;
+  name: TypeName;
   obj: ObjectLiteralExpr;
   range: Range;
 };
@@ -156,11 +171,38 @@ export type StringKeyword = {
   range: Range;
 };
 
-export type TypeRef = {
-  _kind: "TypeRef";
-  name: Identifier;
+export type TypeName = {
+  _kind: "TypeName";
+  name: string;
   range: Range;
 };
+
+export type TypeRef = {
+  _kind: "TypeRef";
+  name: TypeName;
+  range: Range;
+};
+
+export function booleanKeyword(
+  dummy_: string | null = null,
+  range: Range = [0, 0]
+): BooleanKeyword {
+  DEBUG &&
+    (() => {
+      assert(
+        dummy_ === null || typeof dummy_ === "string",
+        `Invalid value for "dummy_" arg in "BooleanKeyword" call.\nExpected: string?\nGot:      ${JSON.stringify(
+          dummy_
+        )}`
+      );
+      assertRange(range, "BooleanKeyword");
+    })();
+  return {
+    _kind: "BooleanKeyword",
+    dummy_,
+    range,
+  };
+}
 
 export function document(
   definitions: Definition[],
@@ -326,15 +368,15 @@ export function objectLiteralExpr(
 }
 
 export function objectTypeDef(
-  name: Identifier,
+  name: TypeName,
   obj: ObjectLiteralExpr,
   range: Range = [0, 0]
 ): ObjectTypeDef {
   DEBUG &&
     (() => {
       assert(
-        name._kind === "Identifier",
-        `Invalid value for "name" arg in "ObjectTypeDef" call.\nExpected: Identifier\nGot:      ${JSON.stringify(
+        name._kind === "TypeName",
+        `Invalid value for "name" arg in "ObjectTypeDef" call.\nExpected: TypeName\nGot:      ${JSON.stringify(
           name
         )}`
       );
@@ -375,12 +417,30 @@ export function stringKeyword(
   };
 }
 
-export function typeRef(name: Identifier, range: Range = [0, 0]): TypeRef {
+export function typeName(name: string, range: Range = [0, 0]): TypeName {
   DEBUG &&
     (() => {
       assert(
-        name._kind === "Identifier",
-        `Invalid value for "name" arg in "TypeRef" call.\nExpected: Identifier\nGot:      ${JSON.stringify(
+        typeof name === "string",
+        `Invalid value for "name" arg in "TypeName" call.\nExpected: string\nGot:      ${JSON.stringify(
+          name
+        )}`
+      );
+      assertRange(range, "TypeName");
+    })();
+  return {
+    _kind: "TypeName",
+    name,
+    range,
+  };
+}
+
+export function typeRef(name: TypeName, range: Range = [0, 0]): TypeRef {
+  DEBUG &&
+    (() => {
+      assert(
+        name._kind === "TypeName",
+        `Invalid value for "name" arg in "TypeRef" call.\nExpected: TypeName\nGot:      ${JSON.stringify(
           name
         )}`
       );
@@ -394,6 +454,7 @@ export function typeRef(name: Identifier, range: Range = [0, 0]): TypeRef {
 }
 
 interface Visitor<TContext> {
+  BooleanKeyword?(node: BooleanKeyword, context: TContext): void;
   Document?(node: Document, context: TContext): void;
   FieldDef?(node: FieldDef, context: TContext): void;
   FloatKeyword?(node: FloatKeyword, context: TContext): void;
@@ -403,6 +464,7 @@ interface Visitor<TContext> {
   ObjectLiteralExpr?(node: ObjectLiteralExpr, context: TContext): void;
   ObjectTypeDef?(node: ObjectTypeDef, context: TContext): void;
   StringKeyword?(node: StringKeyword, context: TContext): void;
+  TypeName?(node: TypeName, context: TContext): void;
   TypeRef?(node: TypeRef, context: TContext): void;
 }
 
@@ -421,6 +483,10 @@ export function visit<TNode extends Node, TContext>(
   context?: TContext
 ): TNode {
   switch (node._kind) {
+    case "BooleanKeyword":
+      visitor.BooleanKeyword?.(node, context);
+      break;
+
     case "Document":
       visitor.Document?.(node, context);
       node.definitions.forEach((d) => visit(d, visitor, context));
@@ -462,6 +528,10 @@ export function visit<TNode extends Node, TContext>(
 
     case "StringKeyword":
       visitor.StringKeyword?.(node, context);
+      break;
+
+    case "TypeName":
+      visitor.TypeName?.(node, context);
       break;
 
     case "TypeRef":