+++ packages/liveblocks-schema/src/checker/__tests__/checker.test.ts
@@ -1,5 +1,0 @@
-describe("checker", () => {
-  test.todo("check good and bad example inputs");
-});
-
-export {};